Fixing the updated groups issue in backend

+++ relestimation/src/main/java/edu/aspen/capstone/estimation/relative/service/ProjectGroupServiceImpl.java
@@ -8,6 +8,7 @@ import edu.aspen.capstone.estimation.relative.utils.JSONExceptionWrapper;
 import edu.aspen.capstone.estimation.relative.utils.JSONResponseWrapper;
 import java.util.ArrayList;
 import java.util.List;
+import org.apache.commons.collections4.CollectionUtils;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -27,7 +28,11 @@ public class ProjectGroupServiceImpl implements ProjectGroupService {
         try {
             ModelMapper modelMapper = new ModelMapper();
             List<ProjectGroups> projects = projectgrpDAO.getAllByProject(prjId);
-            ProjectGroupDO projectGroup = DOUtils.decode(projects);
+            ProjectGroupDO projectGroup = new ProjectGroupDO();
+            projectGroup.setProjectId(prjId);
+            if (CollectionUtils.isNotEmpty(projects)) {
+                projectGroup = DOUtils.decode(projects);
+            }
             return JSONResponseWrapper.getResponseInstance(projectGroup);
         } catch (Exception e) {
             return JSONResponseWrapper.getErrorResponseInstance(
@@ -42,27 +47,37 @@ public class ProjectGroupServiceImpl implements ProjectGroupService {
             List<ProjectGroups> reqList = DOUtils.encode(project);
             List<ProjectGroups> needsUpdate = new ArrayList<ProjectGroups>();
             List<ProjectGroups> needsDelete = new ArrayList<ProjectGroups>();
-            
+
             //let us populate the needs insert&update list, we need to take the reqList 
             //and see if any not present in the current one.
-            for(ProjectGroups grp: currentList){
-                if(reqList.contains(grp)){
-                    needsUpdate.add(grp);
-                } else {
-                    needsDelete.add(grp);
+            if (CollectionUtils.isNotEmpty(currentList)) {
+                for (ProjectGroups grp : currentList) {
+                    if (reqList.contains(grp)) {
+                        needsUpdate.add(grp);
+                    } else {
+                        needsDelete.add(grp);
+                    }
                 }
+
+                for (ProjectGroups grp : reqList) {
+                    if (!currentList.contains(grp)) {
+                        needsUpdate.add(grp);
+                    }
+                }
+            } else {
+                System.out.println("Comming in all added part");
+                needsUpdate.addAll(reqList);
             }
-            
-            for(ProjectGroups grp: reqList){
-                if(!currentList.contains(grp)){
-                    needsUpdate.add(grp);
-                } 
-            }
-            
+
             //Now update insert and update needed and delete the ones we don't
             //need.
-            List<ProjectGroups> updated = projectgrpDAO.updateGroupsForProject(needsUpdate);
-            boolean result = projectgrpDAO.deleteGroupsForProject(needsDelete);
+            if (CollectionUtils.isNotEmpty(needsUpdate)) {
+                System.out.println("calling to updated");
+                List<ProjectGroups> updated = projectgrpDAO.updateGroupsForProject(needsUpdate);
+            }
+            if (CollectionUtils.isNotEmpty(needsDelete)) {
+                boolean result = projectgrpDAO.deleteGroupsForProject(needsDelete);
+            }
             return JSONResponseWrapper.getDefaultSuccessResponseInstance();
         } catch (Exception e) {
             return JSONResponseWrapper.getErrorResponseInstance(new JSONExceptionWrapper("Error", e));
